Implement summary box stats

+++ storage/sorted_test.go
@@ -28,25 +28,25 @@ func TestBasicSortedSet(t *testing.T) {
 	past := time.Now()
 
 	r := db.Retries()
-	assert.Equal(t, 0, r.Size())
+	assert.Equal(t, int64(0), r.Size())
 	err = r.AddElement(util.Thens(past), fmt.Sprintf("0%s", jid), j1)
 	assert.NoError(t, err)
-	assert.Equal(t, 1, r.Size())
+	assert.Equal(t, int64(1), r.Size())
 
 	jid, j2 := fakeJob()
 	err = r.AddElement(util.Thens(past), fmt.Sprintf("1%s", jid), j2)
 	assert.NoError(t, err)
-	assert.Equal(t, 2, r.Size())
+	assert.Equal(t, int64(2), r.Size())
 
 	current := time.Now()
 	jid, j3 := fakeJob()
 	err = r.AddElement(util.Thens(current.Add(10*time.Second)), jid, j3)
 	assert.NoError(t, err)
-	assert.Equal(t, 3, r.Size())
+	assert.Equal(t, int64(3), r.Size())
 
 	results, err := r.RemoveBefore(util.Thens(current.Add(1 * time.Second)))
 	assert.NoError(t, err)
-	assert.Equal(t, 1, r.Size())
+	assert.Equal(t, int64(1), r.Size())
 	assert.Equal(t, 2, len(results))
 	values := [][]byte{j1, j2}
 	assert.Equal(t, values, results)
@@ -60,16 +60,16 @@ func TestRocksSortedSet(b *testing.T) {
 	assert.NoError(b, err)
 	defer db.Close()
 
-	count := 1000
+	count := int64(1000)
 	retries := db.Retries()
 	start := time.Now()
-	for i := 0; i < count; i++ {
+	for i := int64(0); i < count; i++ {
 		jid, job := fakeJob()
-		err = retries.AddElement(util.Thens(start.Add(time.Duration(rand.Intn(10*count))*time.Second)), jid, job)
+		err = retries.AddElement(util.Thens(start.Add(time.Duration(rand.Intn(int(10*count)))*time.Second)), jid, job)
 		assert.NoError(b, err)
 	}
 
-	amt := 0
+	amt := int64(0)
 	ajid := ""
 	atstamp := ""
 	err = retries.EachElement(func(tstamp string, jid string, elm []byte) error {
@@ -78,18 +78,18 @@ func TestRocksSortedSet(b *testing.T) {
 		assert.Equal(b, 27, len(tstamp))
 		assert.Equal(b, 16, len(jid))
 		assert.NotNil(b, elm)
-		amt += 1
+		amt += int64(1)
 		return nil
 	})
 	assert.NoError(b, err)
 	assert.Equal(b, count, amt)
 
-	assert.Equal(b, 0, db.Working().Size())
+	assert.Equal(b, int64(0), db.Working().Size())
 	err = retries.MoveTo(db.Working(), atstamp, ajid, func(payload []byte) (string, []byte, error) {
 		return util.Nows(), payload, nil
 	})
 	assert.NoError(b, err)
-	assert.Equal(b, 1, db.Working().Size())
+	assert.Equal(b, int64(1), db.Working().Size())
 	assert.Equal(b, count-1, retries.Size())
 	count -= 1
 
@@ -98,12 +98,12 @@ func TestRocksSortedSet(b *testing.T) {
 	})
 	assert.Error(b, err)
 
-	remd := 0
+	remd := int64(0)
 	start = time.Now()
-	for i := 0; i < count; i++ {
-		elms, err := retries.RemoveBefore(util.Thens(start.Add(time.Duration(rand.Intn(5*count)) * time.Second)))
+	for i := int64(0); i < count; i++ {
+		elms, err := retries.RemoveBefore(util.Thens(start.Add(time.Duration(rand.Intn(int(5*count))) * time.Second)))
 		assert.NoError(b, err)
-		remd += len(elms)
+		remd += int64(len(elms))
 		assert.Equal(b, count-remd, retries.Size())
 		//assert.True(b, len(elms) == 0 || len(elms) == 1 || len(elms) == 2)
 	}
